+++ teacher.rb
@@ -1,7 +1,9 @@
 require './person'
 
 class Teacher < Person
-  def initialize(specialization, age, name = 'Unknown', parent_permission: true)
+  attr_accessor :specialization
+
+  def initialize(age, specialization, name = 'Unknown', parent_permission: true)
     super(name, age, parent_permission: parent_permission)
     @specialization = specialization
   end
